Extract shared CLI override helpers
