+++ cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI helpers shared between Auto-Coder entry points."""
+
+from .overrides import apply_common_flags, build_overrides
+
+__all__ = ["apply_common_flags", "build_overrides"]
